% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECdata.R
\name{getECdata}
\alias{getECdata}
\title{Extract Environment Canada climate data}
\usage{
getECdata(stations, year_start, year_end, folder, time_frame = c("hourly", "daily", "monthly"), verbose = TRUE, delete = TRUE)
}
\arguments{
\item{year_start}{starting year for data pull}

\item{year_end}{end year for data pull}

\item{folder}{folder path to where you want data saved}

\item{timeframe}{timeframe of the data to pull}

\item{verbose}{include progress bar}

\item{delete}{delete files that failed to download or were corrupted}
}
\value{
dataframe
}
\description{
The function will extract the corresponding climate data
for the supplied supplied weather station IDs during the
given period.
}
\note{
This function requires the other function available in the package "getECurls"
}
\examples{
test <- getECdata(stations = c(52), year_start = 2022, year_end = year(Sys.Date()), folder = "C:/Users/User/Documents/Projects", timeframe = "daily")
}
\references{
https://climate.weather.gc.ca/historical_data/search_historic_data_e.html
https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/
}
\author{
Braeden Klaver
}
