% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECurls.R
\name{getECurls}
\alias{getECurls}
\title{Generate URLs for Environment Canada climate data}
\usage{
getECurls(id, year_start, year_end, timeframe = c("hourly", "daily", "monthly"))
}
\arguments{
\item{id}{A station ID}

\item{year_start}{starting year for data pull}

\item{year_end}{end year for data pull}

\item{timeframe}{timeframe of the data to pull}
}
\value{
list
}
\description{
The function will generate the needed URLs
for the supplied supplied weather station IDs during the
given period.
}
\note{
This function is used in the other function available in the package "getECdata"
}
\examples{
test <- getECurls(id = 52, year_start = 2022, year_end 2023, timeframe = "daily")
}
\references{
https://climate.weather.gc.ca/historical_data/search_historic_data_e.html
https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/
}
\author{
Braeden Klaver
}
