% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECdata.R
\name{getECdata}
\alias{getECdata}
\title{Extract Environment Canada climate data}
\usage{
getECdata(stations, year_start, year_end,
          timeframe = c("hourly", "daily", "monthly"),
          verbose = TRUE)
}
\arguments{
\item{stations}{Vector of weather station IDs to pull for}

\item{year_start}{Starting year for data pull}

\item{year_end}{End year for data pull}

\item{timeframe}{Timeframe of the data to pull, select one of c("hourly", "daily", "monthly")}

\item{verbose}{Include progress bar}
}
\value{
Dataframe
}
\description{
getECdata() returns a data frame with
the corresponding climate data for the supplied supplied
weather station IDs during the given period.
}
\examples{
#An example that  pulls the data into the environment
getECdata(stations = c(52, 55, 200), year_start = 2020, year_end = 2023,
          timeframe = "monthly")
}
\references{
\url{https://climate.weather.gc.ca/historical_data/search_historic_data_e.html}
\url{https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/Station_Inventory_ID_Disclaimer_Metadata_EN.txt}
}
\seealso{
This function wraps the function ECweather::getECurls()
}
\author{
Braeden Klaver
}
