% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECdata.R
\name{getECdata}
\alias{getECdata}
\title{Extract Environment Canada climate data}
\usage{
getECdata(stations, year_start, year_end,
timeframe = c("hourly", "daily", "monthly"),
download = FALSE, folder = NULL, verbose = TRUE, delete = TRUE)
}
\arguments{
\item{stations}{Vector of weather station IDs to pull for}

\item{year_start}{Starting year for data pull}

\item{year_end}{End year for data pull}

\item{timeframe}{Timeframe of the data to pull}

\item{download}{Do you want to download the data to a folder}

\item{folder}{Folder path to where you want data saved}

\item{verbose}{Include progress bar}

\item{delete}{Delete files that failed to download or were corrupted}
}
\value{
Dataframe
}
\description{
getECdata() returns a data frame with
the corresponding climate data for the supplied supplied
weather station IDs during the given period. It will also
save the data to a local folder if download is TRUE.
}
\examples{
getECdata(stations = c(52), year_start = 2022,
year_end = 2023, timeframe = "daily", download = TRUE, folder = getwd())
getECdata(stations = c(52, 55, 200), year_start = 2020,
year_end = 2023, timeframe = "monthly", download = FALSE)
}
\references{
https://climate.weather.gc.ca/historical_data/search_historic_data_e.html
https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/
}
\seealso{
This function wraps the function ECweather::getECurls()
}
\author{
Braeden Klaver
}
