% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECurls.R
\name{getECurls}
\alias{getECurls}
\title{Generate URLs for Environment Canada climate data}
\usage{
getECurls(id, year_start, year_end,
          timeframe = c("hourly", "daily", "monthly"))
}
\arguments{
\item{id}{A station ID}

\item{year_start}{Starting year for data pull}

\item{year_end}{End year for data pull}

\item{timeframe}{Timeframe of the data to pull, select one of c("hourly", "daily", "monthly")}
}
\value{
List
}
\description{
getECurls() generates the URLs
for the supplied weather station IDs during the
given period.
}
\examples{
#An example of pulling URLs for daily data
getECurls(id = 52, year_start = 2022, year_end = 2023, timeframe = "daily")

#An example of pulling URLs for monthly data
getECurls(id = 200, year_start = 2020, year_end = 2023, timeframe = "monthly")
}
\references{
\url{https://climate.weather.gc.ca/historical_data/search_historic_data_e.html}
\url{https://collaboration.cmc.ec.gc.ca/cmc/climate/Get_More_Data_Plus_de_donnees/Station_Inventory_ID_Disclaimer_Metadata_EN.txt}
}
\seealso{
This function is wrapped in the function ECweather::getECdata()
}
\author{
Braeden Klaver
}
